+++ app/sse/index.js
@@ -1,0 +1,5 @@
+const SSE = require("express-sse")
+
+const sse = new SSE()
+
+module.exports = sse
