+++ package.json
@@ -23,6 +23,7 @@
     "crypto-js": "^4.2.0",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
+    "express-sse": "^0.5.3",
     "faker": "^6.6.6",
     "jsonwebtoken": "^9.0.2",
     "moment": "^2.30.1",
